refactor(web): reorganize compare page
Split up the components, also added a "Not found" indication
